requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
redis>=5.0.1
dnspython>=2.5.0
email-validator>=2.1.0.post1
pydantic>=2.9.0
pydantic-settings>=2.1.0
httpx>=0.26.0
pytest>=8.0.0
pytest-asyncio>=0.23.5
respx>=0.20.2
mock>=5.1.0
python-whois>=0.8.0
python-Levenshtein>=0.25.0
prometheus-client>=0.20.0
